refactor(bot): move search and youtube error handler into services
Both directories under utils/music/ are stateful service trees, not
utils. search/ holds SearchEngineManager (a timer) and
ProviderHealthService (a mutable Map keyed by provider);
youtubeErrorHandler/ holds YouTubeErrorAnalyzer behind a
YouTubeErrorHandler facade.

They move together on purpose. search/engineManager and
search/searchContentOnYoutube both import youtubeErrorHandler, so
moving either one alone would leave utils importing services, which is
the inversion #1991 exists to remove rather than relocate.

The external surface is smaller than it first looks. Nothing imports
the search barrel or engineManager from outside; only providerHealth is
reached, by deep path, from ten call sites. An early grep for
'utils/music/search' overcounted badly because it also matches
utils/music/searchQueryCleaner, an unrelated module that does not move.

Depth is unchanged, both trees sit three levels under src/, so the one
outward relative import ('../../../types') and the sibling
'../youtubeErrorHandler' both still resolve untouched.

One reference lived outside src/, a jest.mock() literal in
tests/handlers/player/playerFactory.test.ts, and a src-only scan missed
it until the suite caught it.

decisions/2026-05-16-dependabot-batch-handling-policy.md still names
the old path. Left alone deliberately: it records where the code was in
May, and editing an ADR to match a later move falsifies the record.

Part of #1991. Remaining: titleComparison/service.
