+++ packages/bot/src/services/musicManagement/youtubeErrorHandler/analyzer.spec.ts
