+++ packages/bot/src/services/musicManagement/youtubeErrorHandler/analyzer.ts
