+++ packages/bot/src/services/musicManagement/youtubeErrorHandler/index.ts
