+++ packages/bot/src/services/musicManagement/youtubeErrorHandler/types.ts
